Revert "парсер разнесен на слои"
This reverts commit 12c20ea323e29f74c8e4cd0a2590a5daddd1bd24.
